Fix for Netflix shows disappearing issue (#433)
It appears that in many cases (all in my case) that the Netflix api does not include bookmark or duration data. As a result, the calculatedProgress was generally less than the "Minimum percentage watched" setting, and thus hidden from view.

If the data doesn't exist, we should assume its fully watched so that it can be shown to the user to decide to import.

Co-authored-by: Miles Libbey <mlibbey@apple.com>